+++ Mailer/extensions/mailer-module/src/Tests/Feature/Api/v1/Handlers/Users/SubscribeHandlerTest.php
@@ -6,15 +6,12 @@ namespace Tests\Feature\Api\v1\Handlers\Users;
 use Nette\Http\IResponse;
 use Nette\Utils\Json;
 use Remp\MailerModule\Api\v1\Handlers\Users\SubscribeHandler;
-use Remp\MailerModule\Api\v1\Handlers\Users\UserDeleteApiHandler;
-use Remp\MailerModule\Repositories\UserSubscriptionsRepository;
-use Remp\MailerModule\Repositories\UserSubscriptionVariantsRepository;
 use Tests\Feature\Api\BaseApiHandlerTestCase;
 use Tomaj\NetteApi\Response\JsonApiResponse;
 
 class SubscribeHandlerTest extends BaseApiHandlerTestCase
 {
-    /** @var UserDeleteApiHandler */
+    /** @var SubscribeHandler */
     private $handler;
 
     public function setUp(): void
@@ -44,9 +41,7 @@ class SubscribeHandlerTest extends BaseApiHandlerTestCase
         $this->assertInstanceOf(JsonApiResponse::class, $response);
         $this->assertEquals(IResponse::S200_OK, $response->getCode());
 
-        /** @var UserSubscriptionsRepository $userSubscriptionsRepository */
-        $userSubscriptionsRepository = $this->inject(UserSubscriptionsRepository::class);
-        $isSubscribed = (bool) $userSubscriptionsRepository->getTable()->where([
+        $isSubscribed = (bool) $this->userSubscriptionsRepository->getTable()->where([
             'user_id' => 123,
             'user_email' => 'example@example.com',
             'mail_type_id' => $mailType->id,
@@ -76,9 +71,7 @@ class SubscribeHandlerTest extends BaseApiHandlerTestCase
         $this->assertInstanceOf(JsonApiResponse::class, $response);
         $this->assertEquals(IResponse::S200_OK, $response->getCode());
 
-        /** @var UserSubscriptionsRepository $userSubscriptionsRepository */
-        $userSubscriptionsRepository = $this->inject(UserSubscriptionsRepository::class);
-        $isSubscribed = (bool) $userSubscriptionsRepository->getTable()->where([
+        $isSubscribed = (bool) $this->userSubscriptionsRepository->getTable()->where([
             'user_id' => 123,
             'user_email' => 'example@example.com',
             'mail_type_id' => $mailType->id,
@@ -110,9 +103,7 @@ class SubscribeHandlerTest extends BaseApiHandlerTestCase
         $this->assertInstanceOf(JsonApiResponse::class, $response);
         $this->assertEquals(IResponse::S200_OK, $response->getCode());
 
-        /** @var UserSubscriptionVariantsRepository $userSubscriptionVariantsRepository */
-        $userSubscriptionVariantsRepository = $this->inject(UserSubscriptionVariantsRepository::class);
-        $isSubscribed = (bool) $userSubscriptionVariantsRepository->getTable()->where([
+        $isSubscribed = (bool) $this->userSubscriptionVariantsRepository->getTable()->where([
             'mail_user_subscription.user_id' => 123,
             'mail_user_subscription.user_email' => 'example@example.com',
             'mail_user_subscription.subscribed' => 1,
@@ -144,9 +135,7 @@ class SubscribeHandlerTest extends BaseApiHandlerTestCase
         $this->assertInstanceOf(JsonApiResponse::class, $response);
         $this->assertEquals(IResponse::S200_OK, $response->getCode());
 
-        /** @var UserSubscriptionVariantsRepository $userSubscriptionVariantsRepository */
-        $userSubscriptionVariantsRepository = $this->inject(UserSubscriptionVariantsRepository::class);
-        $isSubscribed = (bool) $userSubscriptionVariantsRepository->getTable()->where([
+        $isSubscribed = (bool) $this->userSubscriptionVariantsRepository->getTable()->where([
             'mail_user_subscription.user_id' => 123,
             'mail_user_subscription.user_email' => 'example@example.com',
             'mail_user_subscription.subscribed' => 1,
@@ -183,9 +172,7 @@ class SubscribeHandlerTest extends BaseApiHandlerTestCase
         $this->assertInstanceOf(JsonApiResponse::class, $response);
         $this->assertEquals(IResponse::S200_OK, $response->getCode());
 
-        /** @var UserSubscriptionsRepository $userSubscriptionsRepository */
-        $userSubscriptionsRepository = $this->inject(UserSubscriptionsRepository::class);
-        $isSubscribed = (bool) $userSubscriptionsRepository->getTable()->where([
+        $isSubscribed = (bool) $this->userSubscriptionsRepository->getTable()->where([
             'user_id' => 123,
             'user_email' => 'example@example.com',
             'mail_type_id' => $mailType->id,
@@ -222,9 +209,7 @@ class SubscribeHandlerTest extends BaseApiHandlerTestCase
         $this->assertInstanceOf(JsonApiResponse::class, $response);
         $this->assertEquals(IResponse::S404_NotFound, $response->getCode());
 
-        /** @var UserSubscriptionVariantsRepository $userSubscriptionVariantsRepository */
-        $userSubscriptionVariantsRepository = $this->inject(UserSubscriptionVariantsRepository::class);
-        $isSubscribed = (bool) $userSubscriptionVariantsRepository->getTable()->where([
+        $isSubscribed = (bool) $this->userSubscriptionVariantsRepository->getTable()->where([
             'mail_user_subscription.user_id' => 123,
             'mail_user_subscription.user_email' => 'example@example.com',
             'mail_user_subscription.subscribed' => 1,