+++ Mailer/extensions/mailer-module/src/Tests/Feature/Api/v1/Handlers/Users/UnsubscribeHandlerTest.php
@@ -6,15 +6,12 @@ namespace Tests\Feature\Api\v1\Handlers\Users;
 use Nette\Http\IResponse;
 use Nette\Utils\Json;
 use Remp\MailerModule\Api\v1\Handlers\Users\UnSubscribeHandler;
-use Remp\MailerModule\Api\v1\Handlers\Users\UserDeleteApiHandler;
-use Remp\MailerModule\Repositories\UserSubscriptionsRepository;
-use Remp\MailerModule\Repositories\UserSubscriptionVariantsRepository;
 use Tests\Feature\Api\BaseApiHandlerTestCase;
 use Tomaj\NetteApi\Response\JsonApiResponse;
 
 class UnsubscribeHandlerTest extends BaseApiHandlerTestCase
 {
-    /** @var UserDeleteApiHandler */
+    /** @var UnSubscribeHandler */
     private $handler;
 
     public function setUp(): void
@@ -44,9 +41,7 @@ class UnsubscribeHandlerTest extends BaseApiHandlerTestCase
         $this->assertInstanceOf(JsonApiResponse::class, $response);
         $this->assertEquals(IResponse::S200_OK, $response->getCode());
 
-        /** @var UserSubscriptionsRepository $userSubscriptionsRepository */
-        $userSubscriptionsRepository = $this->inject(UserSubscriptionsRepository::class);
-        $isNotSubscribed = (bool) $userSubscriptionsRepository->getTable()->where([
+        $isNotSubscribed = (bool) $this->userSubscriptionsRepository->getTable()->where([
             'user_id' => 123,
             'user_email' => 'example@example.com',
             'mail_type_id' => $mailType->id,
@@ -77,9 +72,7 @@ class UnsubscribeHandlerTest extends BaseApiHandlerTestCase
         $this->assertInstanceOf(JsonApiResponse::class, $response);
         $this->assertEquals(IResponse::S200_OK, $response->getCode());
 
-        /** @var UserSubscriptionsRepository $userSubscriptionsRepository */
-        $userSubscriptionsRepository = $this->inject(UserSubscriptionsRepository::class);
-        $isNotSubscribed = (bool) $userSubscriptionsRepository->getTable()->where([
+        $isNotSubscribed = (bool) $this->userSubscriptionsRepository->getTable()->where([
             'user_id' => 123,
             'user_email' => 'example@example.com',
             'mail_type_id' => $mailType->id,
@@ -90,11 +83,7 @@ class UnsubscribeHandlerTest extends BaseApiHandlerTestCase
 
     public function testSuccessfulUnsubscribeWithVariantId()
     {
-        $mailType = $this->createMailTypeWithCategory(
-            "category1",
-            "code1",
-            "name1"
-        );
+        $mailType = $this->createMailTypeWithCategory("category1", "code1", "name1");
         $variant = $this->createMailTypeVariant($mailType, 'Foo', 'foo');
         $this->createMailUserSubscription($mailType, 123, 'example@example.com', $variant->id);
 
@@ -107,19 +96,47 @@ class UnsubscribeHandlerTest extends BaseApiHandlerTestCase
             ])
         ];
 
-        /** @var JsonApiResponse $response */
         $response = $this->handler->handle($params);
         $this->assertInstanceOf(JsonApiResponse::class, $response);
         $this->assertEquals(IResponse::S200_OK, $response->getCode());
 
-        /** @var UserSubscriptionVariantsRepository $userSubscriptionVariantsRepository */
-        $userSubscriptionVariantsRepository = $this->inject(UserSubscriptionVariantsRepository::class);
-        $variantSubscription = $userSubscriptionVariantsRepository->getTable()->where([
+        $this->assertNull($this->userSubscriptionVariantsRepository->getTable()->where([
             'mail_user_subscription.user_id' => 123,
             'mail_user_subscription.user_email' => 'example@example.com',
             'mail_type_variant_id' => $variant->id,
-        ])->fetch();
-        $this->assertNull($variantSubscription);
+        ])->fetch());
+
+        $this->assertTrue($this->userSubscriptionsRepository->isUserUnsubscribed(123, $mailType->id));
+    }
+
+    public function testSuccessfulUnsubscribeWithVariantIdButKeepMailType()
+    {
+        $mailType = $this->createMailTypeWithCategory("category1", "code1", "name1");
+        $variant = $this->createMailTypeVariant($mailType, 'Foo', 'foo');
+        $this->createMailUserSubscription($mailType, 123, 'example@example.com', $variant->id);
+
+        $params = [
+            'raw' => Json::encode([
+                'user_id' => 123,
+                'email' => 'example@example.com',
+                'list_code' => $mailType->code,
+                'variant_id' => $variant->id,
+                'keep_list_subscription' => true,
+            ])
+        ];
+
+        $response = $this->handler->handle($params);
+        $this->assertInstanceOf(JsonApiResponse::class, $response);
+        $this->assertEquals(IResponse::S200_OK, $response->getCode());
+
+        $this->assertNull($this->userSubscriptionVariantsRepository->getTable()->where([
+            'mail_user_subscription.user_id' => 123,
+            'mail_user_subscription.user_email' => 'example@example.com',
+            'mail_type_variant_id' => $variant->id,
+        ])->fetch());
+
+        // user should still be kept in mail type subscription
+        $this->assertFalse($this->userSubscriptionsRepository->isUserUnsubscribed(123, $mailType->id));
     }
 
     public function testSuccessfulUnsubscribeWithVariantCode()
@@ -146,9 +163,7 @@ class UnsubscribeHandlerTest extends BaseApiHandlerTestCase
         $this->assertInstanceOf(JsonApiResponse::class, $response);
         $this->assertEquals(IResponse::S200_OK, $response->getCode());
 
-        /** @var UserSubscriptionVariantsRepository $userSubscriptionVariantsRepository */
-        $userSubscriptionVariantsRepository = $this->inject(UserSubscriptionVariantsRepository::class);
-        $variantSubscription = $userSubscriptionVariantsRepository->getTable()->where([
+        $variantSubscription = $this->userSubscriptionVariantsRepository->getTable()->where([
             'mail_user_subscription.user_id' => 123,
             'mail_user_subscription.user_email' => 'example@example.com',
             'mail_type_variant_id' => $variant->id,
@@ -186,9 +201,7 @@ class UnsubscribeHandlerTest extends BaseApiHandlerTestCase
         $this->assertInstanceOf(JsonApiResponse::class, $response);
         $this->assertEquals(IResponse::S200_OK, $response->getCode());
 
-        /** @var UserSubscriptionsRepository $userSubscriptionsRepository */
-        $userSubscriptionsRepository = $this->inject(UserSubscriptionsRepository::class);
-        $isNotSubscribed = (bool) $userSubscriptionsRepository->getTable()->where([
+        $isNotSubscribed = (bool) $this->userSubscriptionsRepository->getTable()->where([
             'user_id' => 123,
             'user_email' => 'example@example.com',
             'mail_type_id' => $mailType->id,
@@ -225,9 +238,7 @@ class UnsubscribeHandlerTest extends BaseApiHandlerTestCase
         $this->assertInstanceOf(JsonApiResponse::class, $response);
         $this->assertEquals(IResponse::S404_NotFound, $response->getCode());
 
-        /** @var UserSubscriptionVariantsRepository $userSubscriptionVariantsRepository */
-        $userSubscriptionVariantsRepository = $this->inject(UserSubscriptionVariantsRepository::class);
-        $isSubscribed = (bool) $userSubscriptionVariantsRepository->getTable()->where([
+        $isSubscribed = (bool) $this->userSubscriptionVariantsRepository->getTable()->where([
             'mail_user_subscription.user_id' => 123,
             'mail_user_subscription.user_email' => 'example@example.com',
             'mail_user_subscription.subscribed' => 1,