+++ CHANGELOG.md
@@ -14,6 +14,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Added an optimizations for dashboard view to reduce obsolete database calls and speed up the filtering by property. remp/remp#1251
   - Added optional `env` configurations `ARTICLE_TRAFFIC_GRAPH_SHOW_INTERVAL_7D` and `ARTICLE_TRAFFIC_GRAPH_SHOW_INTERVAL_30D` to disable possibility of choosing longer time intervals in the article histogram.
 
+### [Mailer]
+
+- Added `code` attribute in response in `/api/v1/mailers/mail-type-categories` API. remp/crm#2723
+- Added `default_variant_id` attribute in response in `/api/v1/mailers/mail-types` API (same for "v2" and "v3"). remp/crm#2723
+- Added `keep_list_subscription` parameter in `/api/v1/users/un-subscribe` API. remp/crm#2723
+  - By default, when mail type has variants and user unsubscribe from all of them, mail type is automatically unsubscribed too. Parameter `keep_list_subscription` set to `true` changes this behaviour - when last variant is unsubscribed, mail type subscription is retained.   
+- Added API endpoint `/api/v3/mailers/mail-types`, which works similarly to v2, but returns more variant details in API response. remp/crm#2723
+- Added API endpoint `/api/v1/users/is-subscribed` checking if user is subscribed to particular mail type (and variant). remp/crm#2723
+
 ## Archive
 
 - [v2.0](./changelogs/CHANGELOG-v2.0.md)